fix: rename frontend package
