+++ app/frontend/services/camera.py
